envsetup.sh: fix aospremote

+++ envsetup.sh
@@ -1279,12 +1279,17 @@ function godir () {
 
 function aospremote()
 {
+    T=$(gettop)
+    if [ ! "$T" ]; then
+        echo "Couldn't locate the top of the tree.  Try setting TOP." >&2
+        return
+    fi
     git remote rm aosp 2> /dev/null
     if [ ! -d .git ]
     then
         echo .git directory not found. Please run this from the root directory of the Android repository you wish to set up.
     fi
-    PROJECT=`pwd | sed s#$ANDROID_BUILD_TOP/##g`
+    PROJECT=`pwd -P | sed s#$T/##g`
     if (echo $PROJECT | grep -qv "^device")
     then
         PFX="platform/"
